SQL Spatial Query Example
This is an excerpt taken from a much larger and more complex workflow.  The workflow is designed to create a list of mailing addresses and corresponding labels from parcels that intersect 1 or more geometries.  The main part to look at is the 'create polygon string for spatial sql query' sequence in the flowchart.